// packages/oms-wallet/src/generated/waas.gen.ts
export enum GrantKind {
  NativeTransfer = 'nativeTransfer',
  ERC20Transfer = 'erc20Transfer'
}
// ...
export interface NativeTransferGrant {
  to: string
  limit: string
}

export interface ERC20TransferGrant {
  token: string
  to?: string
  limit: string
  cumulative?: boolean
}

export interface Grant {
  kind: GrantKind
  nativeTransfer?: NativeTransferGrant
  erc20Transfer?: ERC20TransferGrant
}

export interface Grants {
  entries: Array<Grant>
}
// ...
export interface SessionInfo {
  sessionId: string
  walletId: string
  signerAddress: string
  grants: Grants
  chainId: string
  expiresAt: string
}

export interface ListSessionsRequest {
  page?: Page
}

export interface ListSessionsResponse {
  sessions: Array<SessionInfo>
  page?: Page
}

export interface GetSessionRequest {
  sessionId: string
}

export interface GetSessionResponse {
  session: SessionInfo
}

export interface GetSessionUsageRequest {
  sessionId: string
  network: string
}

export interface GrantUsage {
  grant: Grant
  used?: string
}